feat(vcd parser):add value change parser

+++ wave-insight-lib/src/data_struct/signal.rs
@@ -4,7 +4,7 @@ use super::CodeLocation;
 #[derive(Debug, PartialEq, Clone)]
 pub struct Signal {
     pub size: usize,
-    pub value_change: Vec<(i32,i32)>,
+    pub value_change: Vec<(i32,i64)>,
     pub same_value_signal: Option<(Vec<String>,String)>,//(module_name,signal_name)
     pub module_path: Vec<String>,//TODO:try to find something instead of String,which looks like hash key
     pub location_define: CodeLocation,//TODO:need file name
